refactored Code  and implemented insertion sort. closes #3, closes #10.

+++ src/Algorithms/CheckIfSorted.js
@@ -1,7 +1,7 @@
 export function checkIfSorted(bars) {
     let flag = true;
     for (let i = 0; i < bars.length-1; i++) {
-        if(parseInt(bars[i].style.height) > parseInt(bars[i+1].style.height)){
+        if(parseFloat(bars[i].style.height) > parseFloat(bars[i+1].style.height)){
             flag = false;
             break;
         }
